[REFACTORING] Split IMAPServerProxyProtocolTest
