+++ server/protocols/protocols-imap4/src/test/java/org/apache/james/imapserver/netty/IMAPServerTest.java
@@ -236,89 +236,6 @@ class IMAPServerTest {
 
 
 
-    @Nested
-    class Proxy {
-        private static final String CLIENT_IP = "255.255.255.254";
-        private static final String PROXY_IP = "255.255.255.255";
-        private static final String RANDOM_IP = "127.0.0.2";
-
-        IMAPServer imapServer;
-        private SocketChannel clientConnection;
-
-        @BeforeEach
-        void beforeEach() throws Exception {
-            imapServer = createImapServer("imapServerProxy.xml");
-            int port = imapServer.getListenAddresses().get(0).getPort();
-            memoryIntegrationResources.getMailboxManager()
-                .createMailbox(MailboxPath.inbox(USER), memoryIntegrationResources.getMailboxManager().createSystemSession(USER));
-            clientConnection = SocketChannel.open();
-            clientConnection.connect(new InetSocketAddress(LOCALHOST_IP, port));
-            readBytes(clientConnection);
-        }
-
-        @AfterEach
-        void tearDown() throws Exception {
-            clientConnection.close();
-            imapServer.destroy();
-        }
-
-        private void addBannedIps(String clientIp) {
-            imapServer.getConnectionChecks().stream()
-                .filter(check -> check instanceof IpConnectionCheck)
-                .map(check -> (IpConnectionCheck) check)
-                .forEach(ipCheck -> ipCheck.setBannedIps(Set.of(clientIp)));
-        }
-
-        @Test
-        void shouldNotFailOnProxyInformation() throws Exception {
-            clientConnection.write(ByteBuffer.wrap(String.format("PROXY %s %s %s %d %d\r\na0 LOGIN %s %s\r\n",
-                "TCP4", CLIENT_IP, PROXY_IP, 65535, 65535,
-                USER.asString(), USER_PASS).getBytes(StandardCharsets.UTF_8)));
-
-            assertThat(new String(readBytes(clientConnection), StandardCharsets.US_ASCII))
-                .startsWith("a0 OK");
-        }
-
-        @Test
-        void shouldDetectAndBanByClientIP() throws IOException {
-            addBannedIps(CLIENT_IP);
-
-            // WHEN connect as CLIENT_IP to PROXY_DESTINATION via PROXY_IP
-            clientConnection.write(ByteBuffer.wrap(String.format("PROXY %s %s %s %d %d\r\na0 LOGIN %s %s\r\n",
-                "TCP4", CLIENT_IP, PROXY_IP, 65535, 65535,
-                USER.asString(), USER_PASS).getBytes(StandardCharsets.UTF_8)));
-
-            // THEN LOGIN should be rejected
-            assertThat(new String(readBytes(clientConnection), StandardCharsets.US_ASCII))
-                .doesNotStartWith("a0 OK");
-        }
-
-        @Test
-        void shouldNotBanByProxyIP() throws IOException {
-            // GIVEN somehow PROXY_IP has been banned by mistake
-            addBannedIps(PROXY_IP);
-
-            clientConnection.write(ByteBuffer.wrap(String.format("PROXY %s %s %s %d %d\r\na0 LOGIN %s %s\r\n",
-                "TCP4", CLIENT_IP, PROXY_IP, 65535, 65535,
-                USER.asString(), USER_PASS).getBytes(StandardCharsets.UTF_8)));
-
-            // THEN CLIENT_IP still can connect
-            assertThat(new String(readBytes(clientConnection), StandardCharsets.US_ASCII))
-                .startsWith("a0 OK");
-        }
-
-        @Test
-        void clientUsageShouldBeNormalWhenClientIPIsNotBanned() throws IOException {
-            addBannedIps(RANDOM_IP);
-
-            clientConnection.write(ByteBuffer.wrap(String.format("PROXY %s %s %s %d %d\r\na0 LOGIN %s %s\r\n",
-                "TCP4", CLIENT_IP, PROXY_IP, 65535, 65535,
-                USER.asString(), USER_PASS).getBytes(StandardCharsets.UTF_8)));
-
-            assertThat(new String(readBytes(clientConnection), StandardCharsets.US_ASCII))
-                .startsWith("a0 OK");
-        }
-    }
 
     @Nested
     class Compress {
